Quantity type fix in ReduceStock.

+++ Products.Application/Products/ReduceStock/ReduceStockCommand.cs
@@ -5,6 +5,6 @@ namespace Products.Application.Products.ReduceStock
     public record ReduceStockCommand : IRequest<bool>
     {
         public int ProductId { get; init; }
-        public int Quantity { get; init; }
+        public short Quantity { get; init; }
     }
 }
